windows/requirements.txt
opencv-python==4.8.1.78
tensorflow==2.13.0
numpy==1.24.3
requests==2.31.0
schedule==1.2.0
pywin32==306
Pillow==10.0.1
pyinstaller==5.13.2
python-dotenv==1.0.0
win10toast==0.9